CSE_106/CSE106-FinalProject-StudyForum-1/requirements.txt
astroid==2.15.4
autopep8==2.0.2
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dill==0.3.6
Flask==2.2.3
Flask-Admin==1.6.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mccabe==0.7.0
mdurl==0.1.2
pbr==5.11.1
platformdirs==3.5.0
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
PyYAML==6.0
requests==2.29.0
rich==13.3.5
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.7
stevedore==5.0.0
tomlkit==0.11.8
typing_extensions==4.5.0
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
